bot/requirements.txt
python-telegram-bot==20.7
httpx==0.25.2
pydantic==2.5.0
python-decouple==3.8
openai==1.3.0
python-dateutil==2.8.2
asyncio==3.4.3
aiofiles==23.2.1